Fix number of inferers considered in the reward cycle (#691)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v           ✰  Thanks for creating a PR! You're awesome! ✰
v Please note that maintainers will only review those PRs with a
completed PR template.
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Purpose of Changes and their Description

`k.lowestInfererScoreEma` was not being updated on a new Epoch. Then,
after the 32 slots get filled up, and when the 33rd inference arrives,
it was checked against the value in `k.lowestInfererScoreEma` from
previous Epoch.

If the new inferer has a greater score, it gets added and the old one
gets removed. But since the old one was from the previous epoch (that
is, the worker in the stored `Score` struct of lowest score), it does
not exist anymore, and the function that removes the inferer won't
return and error if it does not exist. So we don't remove any inferer,
and we add a new one. That's how we end up with 33.

This PR effectively resets that stored lowest score (stored by topic)
upon the start of a new worker submission window. Therefore, whenever
one attempts to move a worker out of the active set, we can be assured
the worker is actually a part of the topic's current active set.

## Link(s) to Ticket(s) or Issue(s) resolved by this PR


https://linear.app/alloralabs/issue/PROTO-2851/fix-number-of-inferers-considered-in-the-reward-cycle

## Are these changes tested and documented?

- [x] If tested, please describe how. If not, why tests are not needed.
   - Unit tests provided
- [x] If documented, please describe where. If not, describe why docs
are not needed.
   - Evident in the code and its comments
- [x] Added to `Unreleased` section of `CHANGELOG.md`?
   - Added to stanza of next release

## Still Left Todo

*Fill this out if this is a Draft PR so others can help.*